Added Skills Section in the profile with a design rework. Added defaults...

See merge request AhmadBerri/erp-advanced-project!10

+++ backend/app/Http/Controllers/ProjectController.php
@@ -2,11 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Assignment;
 use App\Models\Project;
+use App\Models\Team;
+use App\Models\User;
+use DB;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
+use JsonException;
+use Carbon\Carbon;
 
 class ProjectController extends Controller
 {
@@ -16,8 +22,9 @@ class ProjectController extends Controller
      */
     public function index(): JsonResponse
     {
-        $projects = Project::latest()
+        $projects = Project::withCount("teams")
             ->orderBy("status", "ASC")
+            ->latest()
             ->get();
         return response()->json([
             "projects" => $projects,
@@ -26,6 +33,7 @@ class ProjectController extends Controller
 
     /**
      * Create a new project record
+     * @throws JsonException
      */
     public function store(Request $request): JsonResponse
     {
@@ -34,7 +42,21 @@ class ProjectController extends Controller
         ]);
         $inputs["name"] = $request["name"];
         $inputs["slug"] = Str::slug($request["name"], "-");
-        Project::create($inputs);
+        $project = Project::create($inputs);
+        if ($request["teams"]) {
+            $teams = json_decode(
+                $request["teams"],
+                false,
+                512,
+                JSON_THROW_ON_ERROR
+            );
+            $id_teams = [];
+            foreach ($teams as $team) {
+                $id_teams[] = $team->value;
+            }
+            $teams_database = Team::findOrFail($id_teams);
+            $project->teams()->attach($teams_database);
+        }
         return response()->json([
             "message" => "Project Added Successfully",
         ]);
@@ -45,8 +67,117 @@ class ProjectController extends Controller
      */
     public function show(Project $project): JsonResponse
     {
+        $roles = [];
+        $assigned_employees = $project->users;
+        foreach ($assigned_employees as $employee) {
+            $employee->team;
+            foreach ($employee->roles as $role) {
+                if ($role->pivot->project_id === $project->id) {
+                    $roles[] = $role->name;
+                }
+            }
+        }
+        $fetched_teams = DB::table("users")
+            ->select("users.team_id")
+            ->join("teams", "users.team_id", "=", "teams.id")
+            ->join("project_team", "teams.id", "=", "project_team.team_id")
+            ->where("project_team.project_id", $project->id)
+            ->whereNotIn(
+                "users.id",
+                DB::table("assignments")
+                    ->select("user_id")
+                    ->where("assignments.project_id", $project->id)
+                    ->where("assignments.end_date", null)
+                    ->pluck("user_id")
+            )
+            ->distinct()
+            ->pluck("users.team_id");
+        $related_teams = Team::findOrFail($fetched_teams);
         return response()->json([
             "project" => $project,
+            "related_teams" => $related_teams,
+            "assigned_employees" => $assigned_employees,
+            "roles" => $roles,
+        ]);
+    }
+
+    /**
+     * Get A specific teams according to their project
+     * @param Project $project
+     * @return JsonResponse
+     */
+    public function filterByProject(Project $project): JsonResponse
+    {
+        return response()->json([
+            "teams" => Team::whereNotIn(
+                "teams.id",
+                $project
+                    ->teams()
+                    ->pluck("teams.id")
+                    ->toArray()
+            )
+                ->whereNotIn("teams.id", [1, 2])
+                ->get(),
+        ]);
+    }
+
+    /**
+     * Change Status of the project
+     * @param Project $project
+     * @return JsonResponse
+     */
+    public function update_status(Project $project): JsonResponse
+    {
+        $project->status = !$project->status;
+        if ($project->status === true) {
+            $project->finished_at = Carbon::now();
+            $project->save();
+            return response()->json([
+                "message" => "Project Deactivated",
+            ]);
+        }
+        $project->finished_at = null;
+        $project->save();
+        return response()->json([
+            "message" => "Project Activated",
+        ]);
+    }
+
+    /**
+     * Assignment for employees in project
+     * @throws JsonException
+     */
+    public function assign_employees(
+        Project $project,
+        Request $request
+    ): JsonResponse {
+        $request->validate([
+            "assignments" => ["required", "string"],
+        ]);
+        $count = 0;
+        $assignments = json_decode(
+            $request["assignments"],
+            false,
+            512,
+            JSON_THROW_ON_ERROR
+        );
+        $roles = [];
+        foreach ($assignments as $assignment) {
+            $roles[] = $assignment->role;
+        }
+        $employees = [];
+        foreach ($assignments as $assignment) {
+            $employees[] = $assignment->user_id;
+        }
+        $employees_database = User::findOrFail($employees);
+        foreach ($employees_database as $employee) {
+            $project
+                ->users()
+                ->attach($employee->id, ["role_id" => $roles[$count]]);
+            $count++;
+        }
+        return response()->json([
+            "message" => "Assignment was successful",
         ]);
     }
 
@@ -67,8 +198,23 @@ class ProjectController extends Controller
         $inputs["name"] = $request["name"];
         $inputs["slug"] = Str::slug($request["name"], "-");
         $project->update($inputs);
+        if ($request["teams"]) {
+            $teams = json_decode(
+                $request["teams"],
+                false,
+                512,
+                JSON_THROW_ON_ERROR
+            );
+            $id_teams = [];
+            foreach ($teams as $team) {
+                $id_teams[] = $team->value;
+            }
+            $teams_database = Team::findOrFail($id_teams);
+            $project->teams()->attach($teams_database);
+        }
         return response()->json([
             "message" => "Project Successfully Updated",
+            "slug" => $inputs["slug"],
         ]);
     }
 
@@ -79,6 +225,15 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project): JsonResponse
     {
+        if ($project->teams()->exists()) {
+            return response()->json(
+                [
+                    "message" =>
+                        "Project has teams, please remove all teams to delete.",
+                ],
+                403
+            );
+        }
         $project->delete();
         return response()->json([
             "message" => "Project Deleted Successfully",
